Conflicts are partially resolved

+++ satellite/metainfo/service.go
@@ -7,18 +7,24 @@ import (
 	"context"
 	"time"
 
-	"github.com/gogo/protobuf/proto"
-	"github.com/skyrings/skyring-common/tools/uuid"
+	"github.com/zeebo/errs"
 	"go.uber.org/zap"
 
 	"storj.io/common/macaroon"
 	"storj.io/common/pb"
 	"storj.io/common/storj"
+	"storj.io/common/uuid"
+	"storj.io/storj/satellite/metainfo/metabase"
 	"storj.io/storj/storage"
-	"storj.io/uplink/storage/meta"
+	"storj.io/uplink/private/storage/meta"
 )
 
-// Service structure
+var (
+	// ErrBucketNotEmpty is returned when bucket is required to be empty for an operation.
+	ErrBucketNotEmpty = errs.Class("bucket not empty")
+)
+
+// Service provides the metainfo service dependencies.
 //
 // architecture: Service
 type Service struct {
@@ -33,19 +39,43 @@ func NewService(logger *zap.Logger, db PointerDB, bucketsDB BucketsDB) *Service 
 }
 
 // Put puts pointer to db under specific path.
-func (s *Service) Put(ctx context.Context, path string, pointer *pb.Pointer) (err error) {
+func (s *Service) Put(ctx context.Context, key metabase.SegmentKey, pointer *pb.Pointer) (err error) {
 	defer mon.Task()(&ctx)(&err)
+
+	if err := sanityCheckPointer(key, pointer); err != nil {
+		return Error.Wrap(err)
+	}
 
 	// Update the pointer with the creation date
 	pointer.CreationDate = time.Now()
 
-	pointerBytes, err := proto.Marshal(pointer)
+	pointerBytes, err := pb.Marshal(pointer)
 	if err != nil {
 		return Error.Wrap(err)
 	}
 
 	// CompareAndSwap is used instead of Put to avoid overwriting existing pointers
-	err = s.db.CompareAndSwap(ctx, []byte(path), nil, pointerBytes)
+	err = s.db.CompareAndSwap(ctx, storage.Key(key), nil, pointerBytes)
+	return Error.Wrap(err)
+}
+
+// UnsynchronizedPut puts pointer to db under specific path without verifying for existing pointer under the same path.
+func (s *Service) UnsynchronizedPut(ctx context.Context, key metabase.SegmentKey, pointer *pb.Pointer) (err error) {
+	defer mon.Task()(&ctx)(&err)
+
+	if err := sanityCheckPointer(key, pointer); err != nil {
+		return Error.Wrap(err)
+	}
+
+	// Update the pointer with the creation date
+	pointer.CreationDate = time.Now()
+
+	pointerBytes, err := pb.Marshal(pointer)
+	if err != nil {
+		return Error.Wrap(err)
+	}
+
+	err = s.db.Put(ctx, storage.Key(key), pointerBytes)
 	return Error.Wrap(err)
 }
 
@@ -66,8 +96,8 @@ func (s *Service) UnsynchronizedPut(ctx context.Context, path string, pointer *p
 }
 
 // UpdatePieces calls UpdatePiecesCheckDuplicates with checkDuplicates equal to false.
-func (s *Service) UpdatePieces(ctx context.Context, path string, ref *pb.Pointer, toAdd, toRemove []*pb.RemotePiece) (pointer *pb.Pointer, err error) {
-	return s.UpdatePiecesCheckDuplicates(ctx, path, ref, toAdd, toRemove, false)
+func (s *Service) UpdatePieces(ctx context.Context, key metabase.SegmentKey, ref *pb.Pointer, toAdd, toRemove []*pb.RemotePiece) (pointer *pb.Pointer, err error) {
+	return s.UpdatePiecesCheckDuplicates(ctx, key, ref, toAdd, toRemove, false)
 }
 
 // UpdatePiecesCheckDuplicates atomically adds toAdd pieces and removes toRemove pieces from
@@ -80,12 +110,22 @@ func (s *Service) UpdatePieces(ctx context.Context, path string, ref *pb.Pointer
 // Then it will remove the toRemove pieces and then it will add the toAdd pieces.
 // Replacing the node ID and the hash of a piece can be done by adding the
 // piece to both toAdd and toRemove.
-func (s *Service) UpdatePiecesCheckDuplicates(ctx context.Context, path string, ref *pb.Pointer, toAdd, toRemove []*pb.RemotePiece, checkDuplicates bool) (pointer *pb.Pointer, err error) {
+func (s *Service) UpdatePiecesCheckDuplicates(ctx context.Context, key metabase.SegmentKey, ref *pb.Pointer, toAdd, toRemove []*pb.RemotePiece, checkDuplicates bool) (pointer *pb.Pointer, err error) {
 	defer mon.Task()(&ctx)(&err)
+
+	if err := sanityCheckPointer(key, ref); err != nil {
+		return nil, Error.Wrap(err)
+	}
+
+	defer func() {
+		if err == nil {
+			err = sanityCheckPointer(key, pointer)
+		}
+	}()
 
 	for {
 		// read the pointer
-		oldPointerBytes, err := s.db.Get(ctx, []byte(path))
+		oldPointerBytes, err := s.db.Get(ctx, storage.Key(key))
 		if err != nil {
 			if storage.ErrKeyNotFound.Has(err) {
 				err = storj.ErrObjectNotFound.Wrap(err)
@@ -95,7 +135,7 @@ func (s *Service) UpdatePiecesCheckDuplicates(ctx context.Context, path string, 
 
 		// unmarshal the pointer
 		pointer = &pb.Pointer{}
-		err = proto.Unmarshal(oldPointerBytes, pointer)
+		err = pb.Unmarshal(oldPointerBytes, pointer)
 		if err != nil {
 			return nil, Error.Wrap(err)
 		}
@@ -120,7 +160,7 @@ func (s *Service) UpdatePiecesCheckDuplicates(ctx context.Context, path string, 
 			for _, piece := range toAdd {
 				_, ok := nodePieceMap[piece.NodeId]
 				if ok {
-					return nil, ErrNodeAlreadyExists.New("node id already exists in pointer. Path: %s, NodeID: %s", path, piece.NodeId.String())
+					return nil, ErrNodeAlreadyExists.New("node id already exists in pointer. Key: %s, NodeID: %s", key, piece.NodeId.String())
 				}
 				nodePieceMap[piece.NodeId] = struct{}{}
 			}
@@ -162,13 +202,13 @@ func (s *Service) UpdatePiecesCheckDuplicates(ctx context.Context, path string, 
 		pointer.RepairCount = ref.RepairCount
 
 		// marshal the pointer
-		newPointerBytes, err := proto.Marshal(pointer)
+		newPointerBytes, err := pb.Marshal(pointer)
 		if err != nil {
 			return nil, Error.Wrap(err)
 		}
 
 		// write the pointer using compare-and-swap
-		err = s.db.CompareAndSwap(ctx, []byte(path), oldPointerBytes, newPointerBytes)
+		err = s.db.CompareAndSwap(ctx, storage.Key(key), oldPointerBytes, newPointerBytes)
 		if storage.ErrValueChanged.Has(err) {
 			continue
 		}
@@ -183,9 +223,9 @@ func (s *Service) UpdatePiecesCheckDuplicates(ctx context.Context, path string, 
 }
 
 // Get gets decoded pointer from DB.
-func (s *Service) Get(ctx context.Context, path string) (_ *pb.Pointer, err error) {
+func (s *Service) Get(ctx context.Context, key metabase.SegmentKey) (_ *pb.Pointer, err error) {
 	defer mon.Task()(&ctx)(&err)
-	_, pointer, err := s.GetWithBytes(ctx, path)
+	_, pointer, err := s.GetWithBytes(ctx, key)
 	if err != nil {
 		return nil, err
 	}
@@ -193,11 +233,39 @@ func (s *Service) Get(ctx context.Context, path string) (_ *pb.Pointer, err erro
 	return pointer, nil
 }
 
+// GetItems gets decoded pointers from DB.
+// The return value is in the same order as the argument paths.
+func (s *Service) GetItems(ctx context.Context, keys []metabase.SegmentKey) (_ []*pb.Pointer, err error) {
+	defer mon.Task()(&ctx)(&err)
+	storageKeys := make(storage.Keys, len(keys))
+	for i := range keys {
+		storageKeys[i] = storage.Key(keys[i])
+	}
+	pointerBytes, err := s.db.GetAll(ctx, storageKeys)
+	if err != nil {
+		return nil, Error.Wrap(err)
+	}
+
+	pointers := make([]*pb.Pointer, len(pointerBytes))
+	for i, p := range pointerBytes {
+		if p == nil {
+			continue
+		}
+		var pointer pb.Pointer
+		err = pb.Unmarshal([]byte(p), &pointer)
+		if err != nil {
+			return nil, Error.Wrap(err)
+		}
+		pointers[i] = &pointer
+	}
+	return pointers, nil
+}
+
 // GetWithBytes gets the protocol buffers encoded and decoded pointer from the DB.
-func (s *Service) GetWithBytes(ctx context.Context, path string) (pointerBytes []byte, pointer *pb.Pointer, err error) {
+func (s *Service) GetWithBytes(ctx context.Context, key metabase.SegmentKey) (pointerBytes []byte, pointer *pb.Pointer, err error) {
 	defer mon.Task()(&ctx)(&err)
 
-	pointerBytes, err = s.db.Get(ctx, []byte(path))
+	pointerBytes, err = s.db.Get(ctx, storage.Key(key))
 	if err != nil {
 		if storage.ErrKeyNotFound.Has(err) {
 			err = storj.ErrObjectNotFound.Wrap(err)
@@ -206,7 +274,7 @@ func (s *Service) GetWithBytes(ctx context.Context, path string) (pointerBytes [
 	}
 
 	pointer = &pb.Pointer{}
-	err = proto.Unmarshal(pointerBytes, pointer)
+	err = pb.Unmarshal(pointerBytes, pointer)
 	if err != nil {
 		return nil, nil, Error.Wrap(err)
 	}
@@ -214,33 +282,33 @@ func (s *Service) GetWithBytes(ctx context.Context, path string) (pointerBytes [
 	return pointerBytes, pointer, nil
 }
 
-// List returns all Path keys in the pointers bucket
-func (s *Service) List(ctx context.Context, prefix string, startAfter string, recursive bool, limit int32,
+// List returns all Path keys in the pointers bucket.
+func (s *Service) List(ctx context.Context, prefix metabase.SegmentKey, startAfter string, recursive bool, limit int32,
 	metaFlags uint32) (items []*pb.ListResponse_Item, more bool, err error) {
 	defer mon.Task()(&ctx)(&err)
 
 	var prefixKey storage.Key
-	if prefix != "" {
+	if len(prefix) != 0 {
 		prefixKey = storage.Key(prefix)
 		if prefix[len(prefix)-1] != storage.Delimiter {
 			prefixKey = append(prefixKey, storage.Delimiter)
 		}
 	}
 
-	rawItems, more, err := storage.ListV2(ctx, s.db, storage.ListOptions{
+	more, err = storage.ListV2Iterate(ctx, s.db, storage.ListOptions{
 		Prefix:       prefixKey,
 		StartAfter:   storage.Key(startAfter),
 		Recursive:    recursive,
 		Limit:        int(limit),
 		IncludeValue: metaFlags != meta.None,
+	}, func(ctx context.Context, item *storage.ListItem) error {
+		items = append(items, s.createListItem(ctx, *item, metaFlags))
+		return nil
 	})
 	if err != nil {
 		return nil, false, Error.Wrap(err)
 	}
 
-	for _, rawItem := range rawItems {
-		items = append(items, s.createListItem(ctx, rawItem, metaFlags))
-	}
 	return items, more, nil
 }
 
@@ -264,14 +332,14 @@ func (s *Service) createListItem(ctx context.Context, rawItem storage.ListItem, 
 }
 
 // getMetadata adds the metadata to the given item pointer according to the
-// given metaFlags
+// given metaFlags.
 func (s *Service) setMetadata(item *pb.ListResponse_Item, data []byte, metaFlags uint32) (err error) {
 	if metaFlags == meta.None || len(data) == 0 {
 		return nil
 	}
 
 	pr := &pb.Pointer{}
-	err = proto.Unmarshal(data, pr)
+	err = pb.Unmarshal(data, pr)
 	if err != nil {
 		return Error.Wrap(err)
 	}
@@ -296,53 +364,112 @@ func (s *Service) setMetadata(item *pb.ListResponse_Item, data []byte, metaFlags
 }
 
 // Delete deletes a pointer bytes when it matches oldPointerBytes, otherwise it'll fail.
-func (s *Service) Delete(ctx context.Context, path string, oldPointerBytes []byte) (err error) {
+func (s *Service) Delete(ctx context.Context, key metabase.SegmentKey, oldPointerBytes []byte) (err error) {
 	defer mon.Task()(&ctx)(&err)
 
-	err = s.db.CompareAndSwap(ctx, []byte(path), oldPointerBytes, nil)
+	err = s.db.CompareAndSwap(ctx, storage.Key(key), oldPointerBytes, nil)
 	if storage.ErrKeyNotFound.Has(err) {
 		err = storj.ErrObjectNotFound.Wrap(err)
 	}
 	return Error.Wrap(err)
 }
 
-// UnsynchronizedDelete deletes from item from db without verifying whether the pointer has changed in the database.
-func (s *Service) UnsynchronizedDelete(ctx context.Context, path string) (err error) {
+// UnsynchronizedGetDel deletes items from db without verifying whether the pointers have changed in the database,
+// and it returns deleted items.
+func (s *Service) UnsynchronizedGetDel(ctx context.Context, keys []metabase.SegmentKey) ([]metabase.SegmentKey, []*pb.Pointer, error) {
+	storageKeys := make(storage.Keys, len(keys))
+	for i := range keys {
+		storageKeys[i] = storage.Key(keys[i])
+	}
+
+	items, err := s.db.DeleteMultiple(ctx, storageKeys)
+	if err != nil {
+		return nil, nil, Error.Wrap(err)
+	}
+
+	pointerPaths := make([]metabase.SegmentKey, 0, len(items))
+	pointers := make([]*pb.Pointer, 0, len(items))
+
+	for _, item := range items {
+		data := &pb.Pointer{}
+		err = pb.Unmarshal(item.Value, data)
+		if err != nil {
+			return nil, nil, Error.Wrap(err)
+		}
+
+		pointerPaths = append(pointerPaths, metabase.SegmentKey(item.Key))
+		pointers = append(pointers, data)
+	}
+
+	return pointerPaths, pointers, nil
+}
+
+// UnsynchronizedDelete deletes item from db without verifying whether the pointer has changed in the database.
+func (s *Service) UnsynchronizedDelete(ctx context.Context, key metabase.SegmentKey) (err error) {
 	defer mon.Task()(&ctx)(&err)
 
-	err = s.db.Delete(ctx, []byte(path))
+	err = s.db.Delete(ctx, storage.Key(key))
 	if storage.ErrKeyNotFound.Has(err) {
 		err = storj.ErrObjectNotFound.Wrap(err)
 	}
 	return Error.Wrap(err)
 }
 
-// CreateBucket creates a new bucket in the buckets db
+// CreateBucket creates a new bucket in the buckets db.
 func (s *Service) CreateBucket(ctx context.Context, bucket storj.Bucket) (_ storj.Bucket, err error) {
 	defer mon.Task()(&ctx)(&err)
 	return s.bucketsDB.CreateBucket(ctx, bucket)
 }
 
-// GetBucket returns an existing bucket in the buckets db
+// GetBucket returns an existing bucket in the buckets db.
 func (s *Service) GetBucket(ctx context.Context, bucketName []byte, projectID uuid.UUID) (_ storj.Bucket, err error) {
 	defer mon.Task()(&ctx)(&err)
 	return s.bucketsDB.GetBucket(ctx, bucketName, projectID)
 }
 
-// UpdateBucket returns an updated bucket in the buckets db
+// UpdateBucket returns an updated bucket in the buckets db.
 func (s *Service) UpdateBucket(ctx context.Context, bucket storj.Bucket) (_ storj.Bucket, err error) {
 	defer mon.Task()(&ctx)(&err)
 	return s.bucketsDB.UpdateBucket(ctx, bucket)
 }
 
-// DeleteBucket deletes a bucket from the bucekts db
+// DeleteBucket deletes a bucket from the bucekts db.
 func (s *Service) DeleteBucket(ctx context.Context, bucketName []byte, projectID uuid.UUID) (err error) {
 	defer mon.Task()(&ctx)(&err)
+
+	empty, err := s.IsBucketEmpty(ctx, projectID, bucketName)
+	if err != nil {
+		return err
+	}
+	if !empty {
+		return ErrBucketNotEmpty.New("")
+	}
+
 	return s.bucketsDB.DeleteBucket(ctx, bucketName, projectID)
 }
 
-// ListBuckets returns a list of buckets for a project
+// IsBucketEmpty returns whether bucket is empty.
+func (s *Service) IsBucketEmpty(ctx context.Context, projectID uuid.UUID, bucketName []byte) (bool, error) {
+	prefix, err := CreatePath(ctx, projectID, -1, bucketName, []byte{})
+	if err != nil {
+		return false, Error.Wrap(err)
+	}
+
+	items, _, err := s.List(ctx, prefix.Encode(), "", true, 1, 0)
+	if err != nil {
+		return false, Error.Wrap(err)
+	}
+	return len(items) == 0, nil
+}
+
+// ListBuckets returns a list of buckets for a project.
 func (s *Service) ListBuckets(ctx context.Context, projectID uuid.UUID, listOpts storj.BucketListOptions, allowedBuckets macaroon.AllowedBuckets) (bucketList storj.BucketList, err error) {
 	defer mon.Task()(&ctx)(&err)
 	return s.bucketsDB.ListBuckets(ctx, projectID, listOpts, allowedBuckets)
+}
+
+// CountBuckets returns the number of buckets a project currently has.
+func (s *Service) CountBuckets(ctx context.Context, projectID uuid.UUID) (count int, err error) {
+	defer mon.Task()(&ctx)(&err)
+	return s.bucketsDB.CountBuckets(ctx, projectID)
 }
